VSSM: Fix endianness issue in domain properties
The properties for the on_xxx actions must be 16-bit values in
order to avoid failures on big endian systems.

+++ libxkutil/device_parsing.c
@@ -1602,7 +1602,7 @@ static int parse_features(struct domain *dominfo, xmlNode *features)
         return 1;
 }
 
-static void set_action(int *val, xmlNode *child)
+static void set_action(uint16_t *val, xmlNode *child)
 {
         char *action = (char *)xmlNodeGetContent(child);
